move events table to a new component to improve readability (#171)
